requirements.txt
mysql-connector==2.2.9
Flask==1.1.2
mysql-connector-python==8.0.5
werkzeug==0.16.1 
Flask-Cors==3.0.8        
Flask-Login==0.5.0        
flask-oidc==1.4.0        
Flask-RESTful==0.3.8        
flask-restplus==0.13.0
PyYAML==5.3.1
requests==2.22.0
#sudo apt install libmariadb-dev
#sudo apt-get install libmariadbclient-dev
